// src/app/services/enums/floatingiconimageenum.ts
export enum FloatingIconImage {
    CultureGuide = '/assets/img/culture_guide.jpg',
    BrandManual = '/assets/img/brand_manual.jpg',
    Compass = '/assets/img/compass.jpg',
    Tree = '/assets/img/tree.jpg',
    Guardians = '/assets/img/guardians.jpg',
    Magazine = '/assets/img/magazine.jpg',
    Colab = '/assets/img/colab.jpg',
    Us = '/assets/img/us.jpg',
    Stamp = '/assets/img/stamp.jpg',
    Network = '/assets/img/network.jpg',
    Talks = '/assets/img/talk.jpg',
    Gabw = '/assets/img/gabw.jpg',
    Contact = '/assets/img/contact.jpg',
    Timeline = '/assets/img/timeline_icon.jpg'
}